Include Sample_Ids move in NFCore and MaxQuant
